refactor: move stores into sepatate files and some minor changes
