+++ src/util.ts
@@ -1,19 +1,16 @@
+import { AccessTokenClaim } from "./api/apitypes"
+
 // source: https://stackoverflow.com/questions/38552003/how-to-decode-jwt-token-in-javascript-without-using-a-library
 export function parseJwt(token: string): AccessTokenClaim {
-    var base64Url = token.split('.')[1];
-    var base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-    var jsonPayload = decodeURIComponent(atob(base64).split('').map(function (c) {
+    const base64Url = token.split('.')[1];
+    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+    const jsonPayload = decodeURIComponent(atob(base64).split('').map(function (c) {
         return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
     }).join(''));
 
-    return JSON.parse(jsonPayload);
+    const claim = JSON.parse(jsonPayload);
+    return new AccessTokenClaim(claim)
 };
-
-export interface AccessTokenClaim {
-    exp: string
-    scopes: string[]
-    sub: string
-}
 
 export const getScope = "resources:get"
 export const putScope = "resources:put"
